TEMP CHANGE: Bump CHERIoT RTOS source to local branch w/ changes
Includes the latest driver changes to match RTL/HW changes, and also a
bump to the system clock frequency from 30 MHz to 40 MHz. The driver
changes includes updates to the I2C, Ethernet, PWM and GPIO drivers, as
well as board description file updates.

DO NOT COMMIT THIS; THIS CURRENTLY POINTS TO A COMMIT ON MY LOCAL FORK.
IT SHOULD INSTEAD POINT TO A VALID COMMIT ON THE LOWRISC FORK WHEN MY
CHANGES ARE MERGED THERE.