TEMP CHANGE: Bump `sonata-system` submodule source + system clock freq changes
Other changes that are necessary (outside of bumping `cheriot-rtos`) to
bump the system clock frequency for all demos. This includes updating
the relevant compile flag, and also bumping the revision of the
`sonata-system` submodule (used for building the legacy component of the
automotive demo) to the latest main, which now uses a frequency of 40
MHz instead of 30 MHz.

More importantly, bumping the revision of `sonata-system` also lets us
pull in the latest legacy environment changes to match the latest
hardware pin mapping changes, and as such this is required for the
automotive demo's legacy component to continue working.

THIS IS A TEMPORARY COMMIT, BUMPING TO A COMMIT ON MY LOCAL FORK. DO NOT
COMMIT THIS - IT SHOULD INSTEAD POINT TO A MERGED COMMIT ON THE LOWRISC
REPOSITORY.